requirements.txt
# ─── Web API & Data Collection ───────────────
ytmusicapi==1.10.3
mutagen==1.47.0
rapidfuzz==3.13.0

# ─── Audio Feature Extraction ────────────────
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4
scipy==1.13.1

# ─── Machine Learning ────────────────────────
scikit-learn==1.5.2
xgboost==2.1.3
implicit==0.7.2
umap-learn==0.5.7
hdbscan==0.8.40
torch==2.3.1
vowpalwabbit==9.11.2

# ─── Database ────────────────────────────────
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10

# ─── Backend ─────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic==2.10.4
pydantic-settings==2.7.0

# ─── Dashboard ───────────────────────────────
streamlit==1.41.1
plotly==5.24.1
pandas==2.2.3

# ─── Telegram Bot ────────────────────────────
python-telegram-bot==21.9

# ─── Dev & Testing ───────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
black==24.10.0
ruff==0.8.6
